+++ CHANGELOG.md
@@ -18,6 +18,7 @@ In Development
     - Annotate `copyright` line in `docs/conf.py` with `noqa: A001`
     - `test.yml`: Update `actions/checkout` to `v4`
     - `test.yml`: Don't run on pushes to dependabot PR branches
+    - `test.yml`: Cancel concurrent jobs
 
 v2022.10.16
 -----------