+++ CHANGELOG.md
@@ -14,6 +14,7 @@ In Development
 - `pyrepo mkgithub`: Properly sanitize keywords used as topics
 - Use [`ghtoken`](https://github.com/jwodder/ghtoken) for looking up GitHub
   tokens instead of storing them in the pyrepo config file
+- Replace PyYAML with ruamel.yaml
 - Templates:
     - Add 'A' to enabled flake8 checks in `tox.ini`
     - Update coverage paths config in `tox.ini` for Coverage v7